recognise added arg registerInstance
... enabling calling component to access splitInstance.setSizes().
split-view.js : didInsertElement() recognise added arg registerInstance and after createSplit() pass the created splitInstance to registerInstance().
This is for use by Pretzel frontend mapview leftSplit; it will only be required until upgrade of that app Ember 3.22 -> 3.27  which introduces the modifier template helper, enabling instance to be returned by the modifier to the component.